Refactor components
- Organize the component files by category folders
