+++ src/components/common/PlaceholderCard.jsx
